src/rag/embeddings/pyproject.toml
[project]
name = "embeddings-system"
version = "0.1.0"
description = "Flexible embeddings system for emergency healthcare RAG"
authors = [
    {name = "Emergency Healthcare Team", email = "team@example.com"}
]
dependencies = [
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=1.26.4",
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.4",
    "pandas>=2.2.2",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
medical = [
    "biobert-embedding>=0.1.2",
    "scispacy>=0.5.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true